Improve performances of the `StringSearchService` with HTML resources (#162)
`StringSearchService` was sometimes performing very slowly on some HTML resources. The bottleneck was in the parsing from SwiftSoup (third-party dependency). To solve this, we now attempt to parse the resource as strict XML first and fallback on SwiftSoup only when the HTML resource is not a valid XML.

As EPUB resources are supposed to be valid XHTML, this improves performances of EPUB search.